Boss damage hotfix - Patch 0.3
19 / 07 / 2020 - Patch 0.3.1 - Hotfix

Small Hotfix to change the boss damage to make the boss not one shot the player.

Delta:

With this enemy doing so many damage with its main weapon, the number values of the Star Stone effects need to be toned down for a more fair fight for the player.

Star Stone Effects
Power
Damage Amplification
6X -> 1.5x
Heat
Fire Length
3.75
Fire Damage
1